rtsp_server_jetson_device/requirements.txt
pypylon>=3.0.1
opencv-python
PyGObject
PyYAML
numpy
flask
python-socketio
